reverting to 256x256x32,fp32,b4

+++ dataset_dataloader.py
@@ -90,8 +90,8 @@ class LungsDataset(Dataset):
         img = np.expand_dims(img, axis=0)
         mask = np.expand_dims(mask, axis=0)
 
-        img = img.astype(np.float16)
-        mask = mask.astype(np.float16)
+        #img = img.astype(np.float16)
+        #mask = mask.astype(np.float16)
